+++ input.h
@@ -3,8 +3,9 @@
 #include <stdint.h>
 #include <wayland-client.h>
 
-#include "wayland.h"
+#include "cursor-shape.h"
 #include "misc.h"
+#include "wayland.h"
 
 /*
  * Custom defines for mouse wheel left/right buttons.
@@ -33,4 +34,4 @@ void get_current_modifiers(const struct seat *seat,
                            xkb_mod_mask_t *consumed,
                            uint32_t key);
 
-const char *xcursor_for_csd_border(struct terminal *term, int x, int y);
+enum cursor_shape xcursor_for_csd_border(struct terminal *term, int x, int y);