+++ wayland.h
@@ -28,6 +28,7 @@
 #include <fcft/fcft.h>
 #include <tllist.h>
 
+#include "cursor-shape.h"
 #include "fdm.h"
 
 /* Forward declarations */
@@ -145,13 +146,21 @@ struct seat {
     struct {
         uint32_t serial;
 
+        /* Client-side cursor */
         struct wayl_surface surface;
         struct wl_cursor_theme *theme;
         struct wl_cursor *cursor;
+
+        /* Server-side cursor */
+#if defined(HAVE_CURSOR_SHAPE)
+        struct wp_cursor_shape_device_v1 *shape_device;
+#endif
+
         float scale;
         bool hidden;
+        enum cursor_shape shape;
+        char *last_custom_xcursor;
 
-        const char *xcursor;
         struct wl_callback *xcursor_callback;
         bool xcursor_pending;
     } pointer;
@@ -423,6 +432,10 @@ struct wayland {
 
 #if defined(HAVE_XDG_ACTIVATION)
     struct xdg_activation_v1 *xdg_activation;
+#endif
+
+#if defined(HAVE_CURSOR_SHAPE)
+    struct wp_cursor_shape_manager_v1 *cursor_shape_manager;
 #endif
 
     bool presentation_timings;